requirements.txt
flask==2.0.1
pymongo==3.12.0
Werkzeug==2.0.1
requests==2.26.0
python-whois==0.7.3
tld==0.12.6
google==3.0.0
